Finalize THIS refactor
