+++ frontend-next/src/components/app/main_tab/chat.js
@@ -2,7 +2,7 @@ import { Chat, SystemMessage} from "../datatypes"
 import { useState, useEffect } from "react";
 import { Form, useForm } from "react-hook-form";
 import { ref, onValue, set} from "firebase/database";
-import { database } from "../../../src/app/api/firebase-config";
+import { database } from "../../../app/api/firebase-config";
 
 
   // Chatroom Module for Primary Tab
