+++ frontend-next/src/components/app/sidebar/home.js
@@ -1,5 +1,5 @@
 import { Form, useForm } from "react-hook-form";
-import { database } from "../../../src/app/api/firebase-config";
+import { database } from "../../../app/api/firebase-config";
 import { ref, set } from "firebase/database";
 
 // CreateRoom Module for Sidebar Create Tab
